requirements.txt
paho-mqtt===1.6.1
Django==5.0.6
django_cryptography==1.1
djangorestframework==3.15.2
gunicorn==22.0.0
psycopg2-binary==2.9.9
requests==2.32.3
urllib3==1.24.3
python-dotenv==1.0.0
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1
